post-image: Replace post-image by rpi-image package
Add new rpi-image package to generate rpi image and config.txt
instead of post-image script. It allows to do not manually
specify post-image script in custom defconfig, the rpi-image
package is selected by default and will generate rpi image
with config.txt. The only disadvantage here is the image will be
not re-generated if default make target was invoked, so it needs
specify 'rpi-image-reinstall' target in command line.

Signed-off-by: Vadim Kochan <vadim4j@gmail.com>